Fixed test expected result return type

+++ StellaLib.Test/Mathematics/TestCalculation.cs
@@ -10,7 +10,7 @@ namespace StellaLib.Test.Mathematics
         [TestCase(new long[]{1,200,300,400 }    ,ExpectedResult=250)]
         [TestCase(new long[]{-500,-100,-10,50,68},ExpectedResult=-10)]
         [TestCase(new long[]{-1000,-500,-100,-10,50,68},ExpectedResult=-55)]
-        public long Median(long[] values)
+        public double Median(long[] values)
         {
             return Calculation.Median(values);
         }
